refactor: ignore discover error
