+++ pkg/runtime/pluggable/grpc.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpc
+package pluggable
 
 import (
 	"context"
@@ -43,8 +43,8 @@ func instanceIDStreamInterceptor(instanceID string) grpc.StreamClientInterceptor
 	}
 }
 
-// SocketDialer creates a dialer for the given socket.
-func SocketDialer(socket string, additionalOpts ...grpc.DialOption) GRPCConnectionDialer {
+// socketDialer creates a dialer for the given socket.
+func socketDialer(socket string, additionalOpts ...grpc.DialOption) GRPCConnectionDialer {
 	return func(ctx context.Context, name string, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
 		additionalOpts = append(additionalOpts, grpc.WithStreamInterceptor(instanceIDStreamInterceptor(name)), grpc.WithUnaryInterceptor(instanceIDUnaryInterceptor(name)))
 		return SocketDial(ctx, socket, append(additionalOpts, opts...)...)
